extract ConnectDevice screen from send funds flow
