+++ src/const/navigation.js
@@ -139,8 +139,8 @@ export const ScreenName = {
   CosmosDelegationStarted: "CosmosDelegationStarted",
   CosmosDelegationValidator: "CosmosDelegationValidator",
   CosmosDelegationAmount: "CosmosDelegationAmount",
+  CosmosDelegationSelectDevice: "CosmosDelegationSelectDevice",
   CosmosDelegationConnectDevice: "CosmosDelegationConnectDevice",
-  CosmosDelegationValidation: "CosmosDelegationValidation",
   CosmosDelegationValidationError: "CosmosDelegationValidationError",
   CosmosDelegationValidationSuccess: "CosmosDelegationValidationSuccess",
   CosmosRedelegationValidator: "CosmosRedelegationValidator",
