+++ src/screens/ConnectDevice.js
@@ -1,18 +1,20 @@
 // @flow
+import invariant from "invariant";
 import React from "react";
 import { StyleSheet } from "react-native";
+import { useSelector } from "react-redux";
 import SafeAreaView from "react-native-safe-area-view";
+import { getMainAccount } from "@ledgerhq/live-common/lib/account";
 import type {
-  Account,
-  AccountLike,
   Transaction,
   TransactionStatus,
 } from "@ledgerhq/live-common/lib/types";
 import { createAction } from "@ledgerhq/live-common/lib/hw/actions/transaction";
 import connectApp from "@ledgerhq/live-common/lib/hw/connectApp";
 import type { Device } from "@ledgerhq/live-common/lib/hw/actions/types";
-import DeviceAction from "../../components/DeviceAction";
-import { useSignedTxHandler } from "../../logic/screenTransactionHooks";
+import { accountScreenSelector } from "../reducers/accounts";
+import DeviceAction from "../components/DeviceAction";
+import { useSignedTxHandler } from "../logic/screenTransactionHooks";
 
 const action = createAction(connectApp);
 
@@ -25,19 +27,25 @@ type Props = {
 
 type RouteParams = {
   device: Device,
-  account: AccountLike,
-  parentAccount: ?Account,
+  accountId: string,
   transaction: Transaction,
   status: TransactionStatus,
+  context: string,
 };
 
 export default function ConnectDevice({ route }: Props) {
-  const { account, parentAccount, transaction, status } = route.params;
+  const { account: accountLike, parentAccount } = useSelector(
+    accountScreenSelector(route),
+  );
+  invariant(accountLike, "account is required");
+  const account = getMainAccount(accountLike, parentAccount);
+
+  const { transaction, status, context } = route.params;
   const tokenCurrency =
     account.type === "TokenAccount" ? account.token : undefined;
 
   const onResult = useSignedTxHandler({
-    context: "Send",
+    context,
     account,
     parentAccount,
   });
